updated external libs, docker image and native build, use random port during build and tests, openapi client generation and pom.xml refactoring, documentation update
