+++ .gitignore
@@ -1,2 +1,3 @@
 *.log
-target/
+.swagger-codegen/
+target*/
