+++ README.md
@@ -1,5 +1,5 @@
 # zswag-java
-Access & serve zserio services via REST/OpenAPI using **Java** with SpringBoot technology stack.
+Access & serve zserio services via REST/OpenAPI using **Java** with Spring Boot technology stack.
 <br><br>
 
 ## Prerequisites
@@ -7,16 +7,63 @@ Access & serve zserio services via REST/OpenAPI using **Java** with SpringBoot t
    2. Maven 3.6
 <br><br>
 
+## Known issues
+The infrastructure to build native images (executables as well as corresponding Docker images) is present but the built modules do not work at the moment.
+<br><br>
+
+
 ## Usage
 
 ### Build zserio calculator example (including OpenAPI file generation)
+#### Java
 ```bash
-mvn clean verify
+mvn install
+```
+
+or run script
+
+```bash
+build-java.sh
+```
+
+#### Docker images using Java applications
+```bash
+mvn -Pdocker-java spring-boot:build-image
+```
+
+or run script
+
+```bash
+build-docker-java.sh
+```
+
+#### Native image using GraalVM
+```bash
+mvn -Pnative verify
+```
+with GraalVM used as default JDK and native plugin installed
+
+or run script
+
+```bash
+build-native.sh
+```
+with environment variable $GRAALVM_HOME pointing to GraalVM directory and native plugin installed
+
+#### Docker images using native image created with GraalVM
+```bash
+mvn -Pdocker-native spring-boot:build-image
+```
+
+or run script
+
+```bash
+build-docker-native.sh
 ```
 
 ### Generated OpenAPI file during builds
 ```bash
-.../zswag-java/openapi-client/generated-openapi/openapi.yaml
+.../zswag-java/server/generated-openapi/openapi.yaml
 ```
 
 ### Generated OpenAPI client sources during builds
@@ -25,19 +72,98 @@ mvn clean verify
 ```
 
 ### Start REST server
+#### Java (Spring Boot) application
 ```bash
-java -jar server/target/server-0.0.1-SNAPSHOT.jar
+java -jar server/targetJava/server-0.0.1-SNAPSHOT-boot.jar
 # press Ctrl+C to quit the server
 ```
 
-### Start the client console example
+#### Docker image using Java (Spring Boot) application
+Example using docker compose:
+
 ```bash
-java -jar client/target/client-0.0.1-SNAPSHOT.jar
+version: '3'
+
+services:
+  server:
+    image: "zswag.java/server:0.0.1-SNAPSHOT"
+    environment:
+      - JAVA_OPTS=-Dspring.profiles.active="default"
+    volumes:
+      - /tmp:/tmp
+      - /mnt:/mnt:ro
+    ports:
+      - "5000:5000"
+    network_mode: bridge
+    restart: always
+```
+
+Create Docker container and start it:
+
+```bash
+docker-compose up -d
+```
+
+### Start the client console example
+#### Java (Spring Boot) application
+```bash
+java -Dspring.profiles.active="default" -jar client/targetJava/client-0.0.1-SNAPSHOT-boot.jar
+```
+
+#### Docker image using Java (Spring Boot) application
+Addition to docker compose example above:
+
+```bash
+...
+  client:
+    image: "zswag.java/client:0.0.1-SNAPSHOT"
+    depends_on:
+      - "server"
+    environment:
+      - JAVA_OPTS=-Dspring.profiles.active="default" -Dspring.config.additional-location="file:/config/"
+    volumes:
+      - /tmp:/tmp
+      - /mnt:/mnt:ro
+      - <local file path for external config file>:/config
+    network_mode: bridge
+```
+
+In the local config file named **application.yml** you have to override "localhost" with the concrete machine ip address, e.g.
+
+```bash
+feign:
+    url: 192.168.168.1:5000
 ```
 
 ### Start the OpenApi client console example
+#### Java (Spring Boot) application
 ```bash
-java -jar openapi-client/target/openapi-client-0.0.1-SNAPSHOT.jar
+java -Dspring.profiles.active="default" -jar openapi-client/targetJava/openapi-client-0.0.1-SNAPSHOT-boot.jar
+```
+
+#### Docker image using Java (Spring Boot) application
+Addition to docker compose example above:
+
+```bash
+...
+  client:
+    image: "zswag.java/openapi-client:0.0.1-SNAPSHOT"
+    depends_on:
+      - "server"
+    environment:
+      - JAVA_OPTS=-Dspring.profiles.active="default" -Dspring.config.additional-location="file:/config/"
+    volumes:
+      - /tmp:/tmp
+      - /mnt:/mnt:ro
+      - <local file path for external config file>:/config
+    network_mode: bridge
+```
+
+In the local config file named **application.yml** you have to override "localhost" with the concrete machine ip address, e.g.
+
+```bash
+feign:
+    url: 192.168.168.1:5000
 ```
 
 ### URL for OpenAPI Swagger UI
