+++ build-docker-java.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+mvn -Pdocker-java clean spring-boot:build-image $1
