+++ build-docker-native.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+mvn -Pdocker-native clean spring-boot:build-image $1
