+++ build-java.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+mvn -Pjava clean install $1
