--- server/src/test/resources/application.yml
@@ -1,33 +1,0 @@
-feign:
-    client:
-        config:
-            default:
-                connectTimeout: 10000
-                decoder: zswag.java.rest.feign.FeignZserioDecoder
-                encoder: zswag.java.rest.feign.FeignZserioEncoder
-                loggerLevel: basic
-                readTimeout: 10000
-    url: http://localhost:${server.port}
-
-logging:
-    level:
-        root: WARN
-        zswag: INFO
-
-server:
-    port: 5555
-
-spring:
-    application:
-        name: zswag-zserio-calculator-test-server
-
-    main:
-        web-application-type: reactive
-
-    messages:
-        encoding: UTF-8 
-        fallback-to-system-locale: true 
-
-    mvc:
-      async:
-        request-timeout: -1
